requirements.txt
absl-py==0.9.0
appdirs==1.4.4
audioread==2.1.8
barbar==0.2.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
deepspeech==0.9.1
Distance==0.1.3
fastdtw==0.3.4
future==0.18.2
g2p-en==2.1.0
gitdb==4.0.5
GitPython==3.1.0
google-auth==1.15.0
google-auth-oauthlib==0.4.1
grpcio==1.23.0
halo==0.0.31
idna==2.9
importlib-metadata==1.6.0
inflect==4.1.0
iso8601==0.1.13
joblib==0.15.1
kiwisolver==1.2.0
librosa==0.7.2
llvmlite==0.32.1
log-symbols==0.0.14
Markdown==3.2.2
matplotlib==3.2.1
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nltk==3.5
numba==0.45.1
numpy==1.18.4
oauthlib==3.1.0
olefile==0.46
packaging==20.4
paho-mqtt==1.5.1
Pillow==7.1.2
pooch==1.2.0
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytorch-ignite==0.3.0
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
serial==0.0.97
six==1.14.0
smmap==3.0.4
SoundFile==0.10.3.post1
spinners==0.0.24
tensorboard==2.0.0
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
threadpoolctl==2.0.0
torch==1.5.0
torchvision==0.6.0a0+82fd1c8
tornado==6.1
tqdm==4.46.0
Unidecode==1.1.1
urllib3==1.25.9
webrtcvad==2.0.10
Werkzeug==1.0.1
zipp==3.1.0
